Move all the num_workers checking and setting logic to train.py. Its rightful home may be trainer_config.py but the messiness imposed by the ???-value handling is probably not worth it
